#3342 - Macro Action screen enhancements
Added new ability to move the Action Screens Up or Down. Previously
they had to be deleted and re-created in the desired sequence.

Added new ability to Expand Or Collapse the Action Screens. This allows
the user to focus on a specific screen.

The Macro Actions are now automatically collated into the correct
sequence. Input Value actions are sequenced first in Row/Column order,
followed by a Key-Press action.

Added new validation to ensure that the Close Browser/Window action is
the only action on a screen.

Added new validation to ensure that there is only one Press Key action
on a screen.

Added new validation to ensure that multiple Input Value actions cannot
be specified for the same Row/Column.

Added new validation to ensure that Action Data is a mandatory input
field for all applicable Data Types.

Added new ability to Edit an action on the Macro Action Screen, instead
of having to delete it and re-add it.

Added a new Refresh button to the top right of the Genie Macro Action
panel.

Added data consistency checks for Macro Screen and Action data. If
multiple users make a change to the same screen or action, other users
will now get a warning that the data has changed.

Fixed screen selection issue, so a newly added screen will now be
selected after it is added.